feat: add dataset based true classifier
