+++ missingDataTrainingModule/Prediction/classification_network.py
@@ -305,27 +305,3 @@ class ProteinCNN(PredictorAbstract):
         x = self.fc(x)
         return x
 
-
-
-
-classifiers_list = {
-    "none": None,
-    "ClassifierLinear" : ClassifierLinear,
-    "ConvClassifier" : ConvClassifier,
-    "ConvClassifier2" : ConvClassifier2,
-    "ProteinCNN" : ProteinCNN,
-    "ClassifierLvl1" : ClassifierLVL1,
-    "ClassifierLvl2" : ClassifierLVL2,
-    "ClassifierLvl3" : ClassifierLVL3,
-    "RealXClassifier" : RealXClassifier,
-}
-
-
-def get_pred_network(classifier_name):
-    if classifier_name == "none" or classifier_name == None:
-        return None
-    elif classifier_name in classifiers_list:
-        return classifiers_list[classifier_name]
-    else:
-        raise ValueError(f"Classifier {classifier_name} not found")
-
